feat(terminal): complete kitty placeholder decoding

+++ alacritty_terminal/src/graphics/rowcolumn_diacritics.rs
@@ -299,9 +299,9 @@ pub const ROW_COLUMN_DIACRITICS: [char; 297] = [
     '\u{1d244}',
 ];
 
-pub fn diacritic_index(value: char) -> Option<u8> {
+pub fn diacritic_index(value: char) -> Option<u16> {
     ROW_COLUMN_DIACRITICS
         .iter()
         .position(|candidate| *candidate == value)
-        .and_then(|index| u8::try_from(index).ok())
+        .and_then(|index| u16::try_from(index).ok())
 }
